+++ docs/.vitepress/config.ts
@@ -1,0 +1,31 @@
+const guideSidebar = [
+  {
+    text: "Introduction",
+    children: [
+      { text: "What it do?", link: "/" },
+      { text: "Command Line", link: "/guide/cli" },
+    ],
+  },
+  {
+    text: "Advanced",
+    children: [{ text: "API Reference", link: "/guide/api" }],
+  },
+];
+
+export default {
+  title: "@hyrious/esbuild-dev",
+  description: "A simple wrapper of esbuild to run your script file.",
+  themeConfig: {
+    repo: "hyrious/esbuild-dev",
+    docsDir: "docs",
+    editLinks: true,
+    lastUpdated: true,
+
+    nav: [{ text: "Guide", link: "/", activeMatch: "^/$|^/guide/" }],
+
+    sidebar: {
+      "/guide/": guideSidebar,
+      "/": guideSidebar,
+    },
+  },
+};