+++ src/shim.d.ts
@@ -2,3 +2,5 @@ declare module "*.txt" {
   const contents: string;
   export default contents;
 }
+
+declare var __ESM__: boolean;